requirements.txt
# ===================================
# CORE FRAMEWORK - FastAPI & Celery
# ===================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
flower==2.0.1

# ===================================
# DATABASE & ORM
# ===================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# ===================================
# CONFIGURATION & VALIDATION
# ===================================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# ===================================
# WEB SCRAPING - Core
# ===================================
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
aiohttp==3.9.1
httpx==0.25.2

# ===================================
# WEB SCRAPING - Advanced
# ===================================
selenium==4.15.2
playwright==1.40.0
webdriver-manager==4.0.1
scrapy==2.11.0

# ===================================
# FAKE DATA & PROXY
# ===================================
fake-useragent==1.4.0
python-socks[asyncio]==2.3.0

# ===================================
# AI/ML - LLM Integration
# ===================================
ollama==0.1.7
openai==1.3.7
transformers==4.35.2
sentence-transformers==2.2.2

# ===================================
# AI/ML - LangChain
# ===================================
langchain-core==0.1.53
langchain==0.1.12
langchain-community==0.0.38
huggingface-hub==0.17.3

# ===================================
# DATA PROCESSING
# ===================================
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2

# ===================================
# UTILITIES & HELPERS
# ===================================
python-dateutil==2.8.2
tqdm==4.66.1
aiofiles==23.2.1
jinja2==3.1.2

# ===================================
# SECURITY & AUTH
# ===================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===================================
# LOGGING & MONITORING
# ===================================
loguru==0.7.2
structlog==23.2.0
prometheus-client==0.19.0

# ===================================
# TESTING & DEVELOPMENT
# ===================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
responses==0.24.1

# ===================================
# ADDITIONAL SCRAPING TOOLS
# ===================================
newspaper3k==0.2.8
readability-lxml==0.8.1

# ===================================
# TYPE HINTS & COMPATIBILITY
# ===================================
typing-extensions==4.8.0